Admin dialog motoru, X-prefix verb fallback, INPDLG/TRYP, GM verb seti
- 0xAB Gump Value Input + 0xAC reply paketleri eklendi; INPDLG verbi
  artik client'a gercek input prompt aciyor. NetState.GumpTextEntryHandler
  imzasi (serial, context, action, text) Source-X paritesine cevrildi.
- GumpBuilder.AddTextEntryLimited ile DTEXTENTRYLIMITED maxChars
  client'a aktariliyor; INPDLG NAME 30 / BODY 30 / COLOR 16 limitleri
  zorlaniyor.
- ScriptInterpreter'a TRYP <plevel> <verb args...> handler'i eklendi;
  property-edit dialog butonlari (admin tweak / charprop) PLEVEL
  kontrolu ile calisiyor.
- Generic .x<verb> fallback: bilinmeyen X-prefix verbleri target cursor
  acip TrySetProperty -> TryExecuteCommand -> ScriptFallbackExecutor
  zincirinden geciriyor (.xhits, .xkill, .xinvul, .xnuke, .xcolor, ...).
- Yeni GM verbleri: .NUKE/.NUKECHAR/.NUDGE (alan-hedef), .ANIM, .BANK,
  .CONTROL, .DUPE, .HEAL, .SUICIDE, .SUMMONTO, .UNMOUNT, .MOUNT,
  .SUMMONCAGE, .POLY. Tum cevaplar yeni gm_* defmsg'ler ile.
- "guards" / "help guards" handler'i Program.cs'e baglandi: Guarded
  region disinda sessiz, icinde dusman NPC bulunursa "called"
  sysmessage, yoksa "all looks quiet" donuyor.
- Character.cs'te bos kalan verbler dolduruldu: CONSUME (recursive
  backpack walk), ANIM/BOW/SALUTE/SOUND/EFFECT/BARK (BroadcastNearby
  ile cevre clientlara paket), DCLICK/PACK/BANK (owner UI), HUNGRY,
  FIXWEIGHT, DISCONNECT, KICK, RESENDTOOLTIP, INFO, TELE/T, NIGHTSIGHT,
  FOLLOW, SUMMONCAGE, CLEARCTAGS. BOUNCE/DROP no-op (server-side drag
  yok).
- Character.TryGetProperty: MEMORYFINDTYPE.<def>.isValid/.Link.<prop>,
  SKILLCLASS.STATSUM/SKILLSUM/NAME, REPORTEDCLIVER/CLIENTISKR/IS3D/
  ISENHANCED, HOUSES/SHIPS/House.N/Ship.N resolverleri.
- ObjBase: P.X/Y/Z/M chain, TAGCOUNT, TAGAT.N.KEY/VAL iteratoru,
  TAG0./DTAG0. yazma, HUE -> COLOR alias.
- Item: CV_MOVE / CV_FLIP / CV_DCLICK / CV_USE built-in handler'lari.
- GETREFTYPE Source-X bit-mask degerlerine hizalandi (char=0x040000,
  item=0x080000, server=0x000001); RegisterBuiltinDefNames defs.scp
  yuklenmeden once tum [DEFNAME ref_types] degerlerini seed ediyor.
- Server property resolver'a Serv.Map(x,y,z,m).Region.Name parantezli
  fonksiyon, Serv.Resync/Save/Shutdown ve Serv.AllClients
  <fn>/SYSMESSAGE/<verb> dispatch'i eklendi.
- GameClient.ResolveDialogCoord '-N' negatif delta destegi; admin sol
  menusu DORIGIN tabanli yerlesimde dogru konumda. SET / SETUID
  meta-verbleri (Src.set dupe / remove) target cursor ile re-dispatch.

Admin dialog infrastructure, generic X-prefix verb fallback, INPDLG /
TRYP support, and a Source-X-parity GM verb set. Stub verbs on Character
filled in (CONSUME, ANIM, SOUND, EFFECT, DCLICK/PACK/BANK, HUNGRY, ...)
with cross-cutting static delegates wired in Program.cs.

Made-with: Cursor

+++ src/SphereNet.Network/Manager/NetworkManager.cs
@@ -544,7 +544,7 @@ public sealed class NetworkManager : IDisposable
         Action<NetState, uint>? mapDetail = null,
         Action<NetState, uint, byte, byte, ushort, ushort>? mapPinEdit = null,
         // Phase 3
-        Action<NetState, uint, uint, uint, string>? gumpTextEntry = null,
+        Action<NetState, uint, ushort, byte, string>? gumpTextEntry = null,
         Action<NetState, uint>? allNamesRequest = null)
     {
         foreach (var state in _states)
